requirements.txt
arrow==1.3.0
asgiref==3.8.1
binaryornot==0.4.4
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cookiecutter==2.6.0
Django==5.0.2
django-tailwind==4.0.1
gunicorn==22.0.0
icalendar==6.0.0
django-storages==1.14.6
boto3==1.40.18
Pillow==11.3.0
idna==3.10
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
packaging==25.0
psycopg2-binary==2.9.9
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
types-python-dateutil==2.9.0.20250516
urllib3==2.4.0
whitenoise==6.7.0
